+++ packages/components/accordion/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @yamada-ui/accordion
 
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [[`01a717c`](https://github.com/hirotomoyamada/yamada-ui/commit/01a717c877f24bd6052e7a51b5145efec2d29657), [`f63c0f6`](https://github.com/hirotomoyamada/yamada-ui/commit/f63c0f62a9664e3b5bd5afd2ca0702cba32bff00), [`dec603e`](https://github.com/hirotomoyamada/yamada-ui/commit/dec603e919578b7366c21ad51c50e8e744e6f5ca)]:
+  - @yamada-ui/core@1.4.1
+  - @yamada-ui/transitions@1.0.14
+  - @yamada-ui/icon@1.0.13
+  - @yamada-ui/motion@1.0.14
+
 ## 1.0.13
 
 ### Patch Changes