+++ packages/hooks/use-media-query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @yamada-ui/use-media-query
 
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @yamada-ui/providers@1.0.19
+
 ## 1.0.18
 
 ### Patch Changes